requirements.txt
Django==5.2
djangorestframework==3.16.0
PyMySQL==1.1.1
requests~=2.32.3
python-decouple~=3.8
cryptography~=44.0.2
whitenoise
django-cors-headers
routeros-api
sendgrid~=6.11.0